Refactor graph initialization
It's easy to figure out the edges if all nodes are known. For each node,
just check if the left and right node exist in the set of nodes. With
this it's not necessary to construct the graph recursively. Just build
all edges and create a graph based on it.