Optimized minimum path
Instead of figuring an algorithm to get the minimum path faster, changed
the graph to a weighted graph and used the available shortest path
method on all leaves, and got the shortest one. The shortest path
algorithm is O(VlogV + E) so this algorithm is probably fast enough for
large graphs.

Another possible optimization is to break a short path finding if a
minimum was already found and was shorter than the current search.